requirements.txt
absl-py==0.12.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
black==21.4b2
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
detectron2==0.4
easyocr==1.3.1
entrypoints==0.3
future==0.18.2
fvcore==0.1.5.post20210511
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.1
hydra-core==1.1.0rc1
idna==2.10
imageio==2.9.0
importlib-metadata==4.0.1
importlib-resources==5.1.3
iopath==0.1.8
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.2.0
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
layoutparser==0.2.0
Markdown==3.3.4
MarkupSafe==2.0.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numpy==1.20.3
oauthlib==3.1.0
omegaconf==2.1.0rc1
opencv-python==4.5.2.52
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
portalocker==2.3.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytesseract==0.3.7
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.3
Send2Trash==1.5.0
six==1.16.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.9.5
testpath==0.4.4
tifffile==2021.4.8
toml==0.10.2
torch==1.8.1
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.4.1
